+++ src/main/kotlin/omsi/OmsiInstaller.kt
@@ -4,6 +4,7 @@ import dev.nycode.omsilauncher.config.config
 import dev.nycode.omsilauncher.instance.Instance
 import dev.nycode.omsilauncher.instance.LaunchFlag
 import dev.nycode.omsilauncher.util.getOmsiInstallPath
+import dev.nycode.omsilauncher.util.isSteamRunning
 import dev.nycode.omsilauncher.util.logger
 import dev.nycode.omsilauncher.util.parent
 import kotlinx.coroutines.Dispatchers
@@ -85,34 +86,52 @@ private suspend fun doNativeCall(name: String, vararg parameters: String) =
         logger.debug { "Successfully finished native call." }
     }
 
-suspend fun activateInstallationSafe(path: Path) {
+suspend fun activateInstallationSafe(path: Path, startSteam: Boolean = false, awaitSteamDeath: suspend () -> Boolean): Boolean {
     val omsi = getOmsiInstallPath()
     if (omsi.isSymbolicLink() && omsi.readSymbolicLink() == path) {
         val currentManifest = omsi.parent(2) / "appmanifest_$OMSI_STEAM_ID.acf"
         if (currentManifest.exists()) {
             logger.debug { "Backing up $currentManifest to current installation $path" }
             currentManifest.copyTo(path / "manifest.acf", true)
-            return
+            return true
         }
     }
-    doNativeCall("activate-omsi.exe", omsi.absolutePathString(), path.absolutePathString())
+    val isSteamRunning = isSteamRunning()
+
+    if (!isSteamRunning || awaitSteamDeath()) {
+        doNativeCall("activate-omsi.exe", omsi.absolutePathString(), path.absolutePathString())
+
+        if (isSteamRunning && startSteam) {
+            withContext(Dispatchers.IO) {
+                runSteam()
+            }
+        }
+        return true
+    }
+
+    return false
 }
 
-suspend fun activateAndStartInstallationSafe(path: Path, flags: List<LaunchFlag>) {
-    activateInstallationSafe(path)
-    startOmsi(flags)
+suspend fun activateAndStartInstallationSafe(path: Path, flags: List<LaunchFlag>, awaitSteamDeath: suspend () -> Boolean) {
+    if (activateInstallationSafe(path, awaitSteamDeath = awaitSteamDeath)) {
+        startOmsi(flags)
+    }
 }
 
 fun startOmsi(flags: List<LaunchFlag> = emptyList()) {
     val flagsString =
         if (flags.isEmpty()) "" else "// " + flags.joinToString(separator = " ") { it.option }
 
+    runSteam("rungameid/$OMSI_STEAM_ID$flagsString")
+}
+
+private fun runSteam(arguments: String = "") {
     Runtime.getRuntime()
         .exec(
             arrayOf(
                 "rundll32",
                 "url.dll,FileProtocolHandler",
-                "steam://rungameid/$OMSI_STEAM_ID$flagsString"
+                "steam://$arguments"
             )
         )
 }